# swagger/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M&M Car Rental Backend API Documentation",
        "description": "This is an API documentation for M&M Car Rental Backend API. You can find all the endpoints and their descriptions here.",
        "contact": {
            "email": "syahrezafistiferdian32@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "M&M Car Rental Backend Service"
        }
    ],
    "paths": {
...
        "/api/user/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "This endpoint registers a new user by accepting the user's name, email, password, phone number, and address. The system will create a new user record in the database with a default role of 2 (User).",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "phone_number",
                                    "alamat"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Full name of the user",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@gmail.com"
                                    },
                                    "password": {
                                        "description": "User's password (minimum length 6 characters)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "description": "User's phone number",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "alamat": {
                                        "description": "User's address",
                                        "type": "string",
                                        "example": "123 Street Name, City, Country"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the request",
                                            "type": "object",
                                            "example": {
                                                "code": 201,
                                                "is_success": true
                                            }
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "data": {
                                            "description": "Created user data",
                                            "properties": {
                                                "id": {
                                                    "description": "User ID",
                                                    "type": "integer"
                                                },
                                                "nama_user": {
                                                    "description": "User full name",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "User email",
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "description": "User phone number",
                                                    "type": "string"
                                                },
                                                "role_id": {
                                                    "description": "User role ID",
                                                    "type": "integer"
                                                },
                                                "alamat": {
                                                    "description": "User address",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status",
                                            "type": "object",
                                            "example": {
                                                "code": 400,
                                                "is_success": false
                                            }
                                        },
                                        "message": {
                                            "description": "Validation error message",
                                            "type": "string",
                                            "example": "Validation Failed"
                                        },
                                        "data": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status",
                                            "type": "object",
                                            "example": {
                                                "code": 500,
                                                "is_success": false
                                            }
                                        },
                                        "message": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Failed to create user"
                                        },
                                        "data": {
                                            "description": "No additional data",
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "JWT Authorization header using the Bearer scheme.",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}